Create Solution
We can approach this problem using the following steps:

We need to find the maximum possible size of a magic square, which is Math.min(m, n). This is because the side length of the magic square cannot be greater than the width or height of the grid.

We then loop over all side lengths from Math.min(m, n) to 2, and find the largest side length such that there exists at least one magic square with the side length. Once such a side length is found, we return the side length. Otherwise, we return 1.

To find if a magic square of a given side length exists within the grid, we need to check all possible starting positions for the magic square. We do this by looping over all rows and columns where a magic square of the given side length can fit.

For each starting position, we check if the subgrid of the given side length is a magic square. To do this, we calculate the row sums, column sums, and diagonal sums of the subgrid, and check if they are all equal. If they are, we return true.

If we have checked all possible starting positions and none of them are magic squares, we return false.